Demo update 9
Alright Liam, the models have been imported to the unreal demo. Sadly, I had to do it in my branch as when I tried opening the demo with the master branch, it kept crashing. sorry if this effects your workload, but it was the only way to complete the tesk