Dynamic day and night cycle
Instead of using the level blueprint, I made a blueprint which does the same job but it's more easier when for multiple levels. Put a moon which will be at night with a low moon light, the day will last 10 minutes in real time instead of 24 seconds. Both the moon and sun now move not just up and down but east to west.