Dynamic day and night cycle
Added a moon to the sky which has been copied from the actual content of the engine version (4.16) and then edited by me. Both moon and sun now moves like it should in real life. The day lasts for 10 minutes in real time and has a much longer night time as well. Changed the delay of depletion of the hunger to last the first 10 mins. When the moon or sun is below the ground, changed the intensity of the light to zero and then back to normal once it's time.