new map update
Alright, I was able to add all the doors for the new level, though sadly not all the houses have furiture. But worry not, I have marked a house with all its furiture in with a red oil tank with an axe enbeded in to it so you will know where to find it.